+++ tryon/api/pruna/__init__.py
@@ -8,6 +8,7 @@ the same client under the hood.
 
 from .p_image import PImageAdapter
 from .p_image_edit import PImageEditAdapter
+from .p_image_ideogram import PImageIdeogramAdapter
 from .p_image_upscale import PImageUpscaleAdapter
 from .p_video import PVideoAdapter
 from .p_video_animate import PVideoAnimateAdapter
@@ -17,6 +18,7 @@ from .p_video_replace import PVideoReplaceAdapter
 __all__ = [
     "PImageAdapter",
     "PImageEditAdapter",
+    "PImageIdeogramAdapter",
     "PImageUpscaleAdapter",
     "PVideoAdapter",
     "PVideoAnimateAdapter",